Refactor: Push resource list enrichment into ResourcesRepository
Moves the logic for looking up and joining ratings and tags with
libraries out of `ResourcesViewModel` and into a new
`getEnrichedLibraries` method in `ResourcesRepository`. This new
method returns a domain-level `LibraryWithMetadata` data class,
maintaining the layered architecture by keeping presentation-level
mapping in the ViewModel. Also removes a redundant `!it.isPrivate`
check.
