+++ app/src/main/java/org/ole/planet/myplanet/ui/resources/ResourcesViewModel.kt
@@ -10,8 +10,10 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import org.ole.planet.myplanet.MainApplication.Companion.isServerReachable
 import org.ole.planet.myplanet.callback.OnSyncListener
+import org.ole.planet.myplanet.model.ResourceItem
 import org.ole.planet.myplanet.model.ResourceListModel
 import org.ole.planet.myplanet.model.SyncState
+import org.ole.planet.myplanet.model.TagItem
 import org.ole.planet.myplanet.repository.ResourcesRepository
 import org.ole.planet.myplanet.services.SharedPrefManager
 import org.ole.planet.myplanet.services.sync.ServerUrlMapper
@@ -68,6 +70,24 @@ class ResourcesViewModel @Inject constructor(
     }
 
     suspend fun getLibraryListModels(isMyCourseLib: Boolean, modelId: String?): List<ResourceListModel> {
-        return resourcesRepository.getEnrichedLibraryListModels(isMyCourseLib, modelId)
+        val enrichedLibraries = resourcesRepository.getEnrichedLibraries(isMyCourseLib, modelId)
+
+        return enrichedLibraries.map { (library, rating, libraryTags) ->
+            val item = ResourceItem(
+                id = library.id,
+                title = library.title,
+                description = library.description,
+                createdDate = library.createdDate,
+                averageRating = library.averageRating,
+                timesRated = library.timesRated,
+                resourceId = library.resourceId,
+                isOffline = library.isResourceOffline(),
+                _rev = library._rev,
+                uploadDate = library.uploadDate,
+                filename = library.filename
+            )
+            val tags = libraryTags.map { tag -> TagItem(tag.id, tag.name) }
+            ResourceListModel(library, item, rating, tags)
+        }
     }
 }
